moved 'widgets' to a different folder
