+++ packages/tools/src/utilities/voi/colorbar/ColorbarCanvas.ts
@@ -1,6 +1,6 @@
 import { IColorMapPreset } from '@kitware/vtk.js/Rendering/Core/ColorTransferFunction/ColorMaps';
 import { utilities } from '@cornerstonejs/core';
-import interpolateVec3 from '../../../math/vec3/interpolateVec3';
+import interpolateVec3 from '../../math/vec3/interpolateVec3';
 import { ColorbarCanvasProps } from './types/ColorbarCanvasProps';
 import type { ColorbarImageRange, ColorbarVOIRange } from './types';
 import type { ColorbarSize } from './types/ColorbarSize';
