+++ packages/tools/src/utilities/voi/colorbar/ViewportColorbar.ts
@@ -9,7 +9,7 @@ import {
 } from '@cornerstonejs/core';
 import { Colorbar } from './Colorbar';
 import type { ViewportColorbarProps, ColorbarVOIRange } from './types';
-import { getVOIMultipliers } from '../../../getVOIMultipliers';
+import { getVOIMultipliers } from '../../getVOIMultipliers';
 
 const { Events } = Enums;
 const defaultImageRange = { lower: -1000, upper: 1000 };
