+++ packages/tools/src/utilities/voi/colorbar/common/areColorbarSizesEqual.ts
