+++ packages/tools/src/utilities/voi/colorbar/common/isColorbarSizeValid.ts
