+++ packages/tools/src/utilities/voi/colorbar/common/isRangeTextPositionValid.ts
