+++ packages/tools/src/utilities/voi/colorbar/common/isRangeValid.ts
