+++ packages/tools/src/utilities/voi/colorbar/enums/ColorbarRangeTextPosition.ts
