+++ packages/tools/src/utilities/voi/colorbar/types/ColorbarCommonProps.ts
