+++ packages/tools/src/utilities/voi/colorbar/types/ColorbarImageRange.ts
