+++ packages/tools/src/utilities/voi/colorbar/types/ColorbarTicksProps.ts
