+++ packages/tools/src/utilities/voi/colorbar/types/ColorbarTicksStyle.ts
