--- packages/tools/src/utilities/voi/widgets/index.ts
@@ -1,1 +1,0 @@
-export * as colorbar from './colorbar';
