+++ packages/tools/src/widgets/types/WidgetProps.ts
